+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Removed all pre-defined images from the library to escape unbounded maintenance work.
   See https://github.com/testcontainers/testcontainers-rs/issues/471 for details.
+- Removed explicit support for podman.
+  See https://github.com/testcontainers/testcontainers-rs/issues/422 for details.
 
 ## [0.14.0] - 2022-05-30